+++ packages/app-admin-ui/client/classes/List.js
@@ -9,7 +9,7 @@ export const gqlCountQueries = lists => gql`{
 export default class List {
   constructor(
     { access, adminConfig, adminDoc, fields, gqlNames, key, label, path, plural, singular },
-    adminMeta,
+    { readViews, preloadViews, getListByKey, adminPath, authStrategy },
     views
   ) {
     this.access = access;
@@ -21,10 +21,15 @@ export default class List {
     this.path = path;
     this.plural = plural;
     this.singular = singular;
+    this.fullPath = `${adminPath}/${path}`;
 
     this.fields = fields.map(fieldConfig => {
-      const [Controller] = adminMeta.readViews([views[fieldConfig.path].Controller]);
-      return new Controller(fieldConfig, adminMeta, views[fieldConfig.path]);
+      const [Controller] = readViews([views[fieldConfig.path].Controller]);
+      return new Controller(
+        fieldConfig,
+        { readViews, preloadViews, getListByKey, adminPath, authStrategy },
+        views[fieldConfig.path]
+      );
     });
 
     this._fieldsByPath = arrayToObject(this.fields, 'path');