+++ packages/fields/src/types/Relationship/views/Field.js
@@ -12,7 +12,7 @@ import { IconButton } from '@arch-ui/button';
 import Tooltip from '@arch-ui/tooltip';
 
 import RelationshipSelect from './RelationshipSelect';
-import { ListProvider, useList } from '@keystonejs/app-admin-ui/components';
+import { CreateItemModal, ListProvider, useList } from '@keystonejs/app-admin-ui/components';
 
 const MAX_IDS_IN_FILTER = 100;
 
@@ -57,7 +57,7 @@ function SetAsCurrentUser({ listKey, value, onAddUser, many }) {
 
 function LinkToRelatedItems({ field, value }) {
   const { many } = field.config;
-  const { adminPath } = field.adminMeta;
+  const { adminPath } = field;
   const { path } = field.getRefList();
   let isDisabled = false;
   let label;
@@ -103,7 +103,7 @@ function LinkToRelatedItems({ field, value }) {
   );
 }
 
-function CreateAndAddItem({ field, item, onCreate, CreateItemModal }) {
+function CreateAndAddItem({ field, item, onCreate }) {
   const { list, openCreateItemModal } = useList();
 
   let relatedList = field.getRefList();
@@ -167,7 +167,6 @@ const RelationshipField = ({
   onChange,
   item,
   list,
-  CreateItemModal,
 }) => {
   const handleChange = option => {
     const { many } = field.config;
@@ -179,7 +178,7 @@ const RelationshipField = ({
   };
 
   const { many, ref } = field.config;
-  const { authStrategy } = field.adminMeta;
+  const { authStrategy } = field;
   const htmlID = `ks-input-${field.path}`;
 
   const relatedList = field.getRefList();
@@ -209,7 +208,6 @@ const RelationshipField = ({
             field={field}
             item={item}
             list={list}
-            CreateItemModal={CreateItemModal}
           />
         </ListProvider>
         {authStrategy && ref === authStrategy.listKey && (