+++ packages/fields/src/types/Virtual/Implementation.js
@@ -4,18 +4,25 @@ import { KnexFieldAdapter } from '@keystonejs/adapter-knex';
 import { parseFieldAccess } from '@keystonejs/access-control';
 
 export class Virtual extends Implementation {
-  constructor() {
+  constructor(
+    path,
+    { resolver, graphQLReturnType = 'String', graphQLReturnFragment = '', extendGraphQLTypes = [] }
+  ) {
     super(...arguments);
+    this.resolver = resolver;
+    this.graphQLReturnType = graphQLReturnType;
+    this.graphQLReturnFragment = graphQLReturnFragment;
+    this.extendGraphQLTypes = extendGraphQLTypes;
   }
 
   gqlOutputFields() {
-    return [`${this.path}: ${this.config.graphQLReturnType || `String`}`];
+    return [`${this.path}: ${this.graphQLReturnType}`];
   }
   getGqlAuxTypes() {
-    return this.config.extendGraphQLTypes || [];
+    return this.extendGraphQLTypes;
   }
   gqlOutputFieldResolvers() {
-    return { [`${this.path}`]: this.config.resolver };
+    return { [`${this.path}`]: this.resolver };
   }
   gqlQueryInputFields() {
     return [];
@@ -23,8 +30,7 @@ export class Virtual extends Implementation {
   extendAdminMeta(meta) {
     return {
       ...meta,
-      isOrderable: false,
-      graphQLSelection: this.config.graphQLReturnFragment || '',
+      graphQLSelection: this.graphQLReturnFragment,
       isReadOnly: true,
     };
   }