+++ packages/mongo-join-builder/lib/query-parser.js
@@ -33,7 +33,7 @@ function queryParser({ listAdapter, getUID }, query, pathSoFar = [], include) {
         ),
         { AND: '$and', OR: '$or' }[key]
       );
-    } else if (['$search', '$orderBy', '$skip', '$first', '$count'].includes(key)) {
+    } else if (['$search', '$sortBy', '$orderBy', '$skip', '$first', '$count'].includes(key)) {
       return { postJoinPipeline: [modifierTokenizer(listAdapter, query, key, path)] };
     } else if (key === 'id') {
       if (getType(value) === 'Object') {